Rework all stencils and jit
The previous version used more or less random data for the ssas and slots arrays,
and we then patched the addresses of elements of those array into the
stencils.
This was problematic, because when trying to call the resulting jitted functions
with proper arguments, we could not change the values behind those hardcoded addresses,
because the jl_value_t *s are opaque pointers
(or rather should not, although I don't even know how you would patch
them).

This version fixes this by adding one more indirection layer for values
to stencils. With this we can really use place holder pointers when jitting
and later on replace them with addresses of the caller's arguments.